Move organization category operations into ConfigurationController
