+++ test/CareTogether.Core.Test/CareTogether.Core.Test.csproj
@@ -16,6 +16,7 @@
     <PackageReference Include="Moq" Version="4.20.72" />
     <PackageReference Include="MSTest.TestAdapter" Version="4.2.2" />
     <PackageReference Include="MSTest.TestFramework" Version="4.2.2" />
+    <PackageReference Include="Microsoft.FeatureManagement.AspNetCore" Version="4.5.0" />
     <PackageReference Include="coverlet.collector" Version="10.0.0">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
@@ -23,6 +24,11 @@
   </ItemGroup>
 
   <ItemGroup>
+    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\src\CareTogether.Api\CareTogether.Api.csproj" />
     <ProjectReference Include="..\..\src\CareTogether.Core\CareTogether.Core.csproj" />
     <ProjectReference Include="..\CareTogether.TestData\CareTogether.TestData.csproj" />
   </ItemGroup>
